Thornhill/requirements.txt
Django==1.10.3
docutils==0.12
gunicorn==19.6.0
South==1.0.2
Unipath==1.1
django-crispy-forms==1.6.1
celery[redis]
djangorestframework=3.5.3
